+++ history.md
@@ -1,10 +1,12 @@
+* Added styles for MenuWidget; new default style
+* Haywire status working again
 * Combat won't start before START trigger checked
 * Synth monsters now have treasure recoverable with Biotechnology skill
 * TeamRetreatAlert is working now
 * Mission on_win, on_loss methods will not be called more than once each
 * Color editor uses correct channels for mecha
 
-v1.005alpha February 27, 2026
+v1.005alpha March 4, 2026
 * NPC mecha selected based on the NPC's skills
 * Random mecha units now properly use roles
 * Fixed NPCs using repair skill on empty tile bug